feat(everywhere): use es6 features  - ensure use of es6 modules [Delivers #165399662]
