+++ .babelrc
@@ -1,9 +1,3 @@
 {
-    "presets": [
-        ["env", {
-            "targets": {
-              "node": "current"
-            }
-        }]
-    ]
+    "presets": ["@babel/preset-env"]
 }
