--- bin/www
@@ -1,10 +1,0 @@
-
-var debug = require('debug')('Express4');
-var app = require('../app');
-
-app.set('port', process.env.PORT || 3000);
-
-var server = app.listen(app.get('port'), function() {
-  debug('Express server listening on port ' + server.address().port);
- console.log('Express server listening on localhost:3000');
-});
