+++ package.json
@@ -3,8 +3,10 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www",
-    "test": "nyc --require babel-core/register mocha test/test.js --exit",
+    "start": "nodemon --exec babel-node server/app.js",
+    "build": "babel server --out-dir dist",
+    "serve": "node dist/app.js",
+    "test": "nyc mocha test/test.js --exit",
     "coverage": "nyc npm test && nyc report --reporter=text-lcov | coveralls"
   },
   "engines": {
@@ -12,9 +14,6 @@
     "npm": "~6.4.1"
   },
   "dependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-preset-env": "^1.7.0",
     "bcrypt": "^3.0.5",
     "body-parser": "~1.8.1",
     "chai": "^4.2.0",
@@ -30,11 +29,14 @@
     "mocha": "^6.1.2",
     "mocha-lcov-reporter": "^1.3.0",
     "morgan": "~1.3.0",
-    "passport": "^0.2.2",
-    "passport-local": "^1.0.0",
+    "nodemon": "^1.18.11",
     "request": "~2.51.0"
   },
   "devDependencies": {
+    "@babel/cli": "^7.4.3",
+    "@babel/core": "^7.4.3",
+    "@babel/node": "^7.2.2",
+    "@babel/preset-env": "^7.4.3",
     "istanbul": "^0.4.5",
     "nyc": "^13.3.0"
   }
